fix(library): split approval-status server helpers from client bundle
LibrarySurface is a client component that imported formatting helpers
from approval-status.ts, which also pulled in @/lib/db and the full
server-only env chain. Move DB accessors to approval-status.server.ts
so the client module stays boundary-safe.
